Null-merge branch 5.5 (commit ac05f5b) into 5.6

+++ sql/log_event.cc
@@ -1366,7 +1366,7 @@ Log_event* Log_event::read_log_event(IO_CACHE* file,
   Log_event *res=  0;
 #ifndef max_allowed_packet
   THD *thd=current_thd;
-  uint max_allowed_packet= thd ? slave_max_allowed_packet : ~0U;
+  ulong max_allowed_packet= thd ? slave_max_allowed_packet : ~(ulong)0;
 #endif
 
   ulong const max_size=
